+++ RELEASENOTES.md
@@ -1,13 +1,13 @@
 These are the release notes for the [mediawiki-api-base](README.md).
 
-## Version 2.0.0
+## Version 2.0.0 (18 December 2015)
 * Added `MediawikiApi::newFromApiEndpoint` and `MediawikiApi::newFromPage`
 * MediawikiApi constructor access marked as private (please use static factory methods)
 * Added async methods to MediawikiApi `getRequestAsync` & `postRequestAsync`
 * Requires "guzzlehttp/guzzle": "~6.0" ( From "guzzle/guzzle": "~5.2" )
 * Requires "guzzlehttp/promises": "~1.0"
 
-## Version 1.1.0 (5 september 2015)
+## Version 1.1.0 (5 September 2015)
 * Requests that encounter a connection exception are now retried
 * Requests that result in non blocking mediawiki api error codes are now retried (ratelimited, readonly, internal_api_error_DBQueryError)
 * MediawikiApi now implements PSR-3 LoggerAwareInterface